requirements.txt
streamlit
torch
torchvision
Pillow
numpy
scikit-image
tqdm
